% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vips-.R
\name{as.character.Vips}
\alias{as.character.Vips}
\title{Character representation of Vips object}
\usage{
\method{as.character}{Vips}(x, ...)
}
\arguments{
\item{x}{Vips object}

\item{...}{other arguments ignored}
}
\description{
Character representation of Vips object
}
