% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{as_cli_string}
\alias{as_cli_string}
\alias{as_cli_string.numeric}
\alias{as_cli_string.character}
\alias{as_cli_string.logical}
\title{Convert an object to a character string for the command line}
\usage{
as_cli_string(x, ...)

\method{as_cli_string}{numeric}(x, ...)

\method{as_cli_string}{character}(x, ...)

\method{as_cli_string}{logical}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other arguments ignored}
}
\value{
a representation of the object for the command line.
}
\description{
Convert an object to a character string for the command line
}
