% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vips-.R
\name{vips_array}
\alias{vips_array}
\title{Convert an array of values into a single space-separated string}
\usage{
vips_array(...)
}
\arguments{
\item{...}{values to preparse a vips array argument}
}
\value{
single character string with all the original vector elements separated by spaces
}
\description{
Convert an array of values into a single space-separated string
}
