% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vips-.R
\name{vips_array_files}
\alias{vips_array_files}
\title{Convert an array of filenames into a single space-separated string}
\usage{
vips_array_files(...)
}
\arguments{
\item{...}{values to preparse a vips array argument}
}
\value{
single character string with all the original vector elements separated by spaces
}
\description{
Compared to \code{vips_array}, this function tries to escape characters
(such as space) in order to be correctly handled by the shell
}
